+++ server/src/test/java/org/allaymc/api/block/data/BlockFaceTest.java
@@ -1,6 +1,6 @@
 package org.allaymc.api.block.data;
 
-import org.joml.Vector3f;
+import org.joml.Vector3d;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -26,12 +26,12 @@ class BlockFaceTest {
      */
     @Test
     void testRotateVector() {
-        var vec = new Vector3f(0.1f, 0.2f, 0.3f);
+        var vec = new Vector3d(0.1, 0.2, 0.3);
         assertEquals(vec.toString(), BlockFace.EAST.rotateVector(vec).toString());
-        assertEquals(new Vector3f(0.2f, 0.9f, 0.3f).toString(), BlockFace.DOWN.rotateVector(vec).toString());
-        assertEquals(new Vector3f(0.8f, 0.1f, 0.3f).toString(), BlockFace.UP.rotateVector(vec).toString());
-        assertEquals(new Vector3f(0.9f, 0.2f, 0.7f).toString(), BlockFace.WEST.rotateVector(vec).toString());
-        assertEquals(new Vector3f(0.7f, 0.2f, 0.1f).toString(), BlockFace.SOUTH.rotateVector(vec).toString());
-        assertEquals(new Vector3f(0.3f, 0.2f, 0.9f).toString(), BlockFace.NORTH.rotateVector(vec).toString());
+        assertEquals(new Vector3d(0.2, 0.9, 0.3).toString(), BlockFace.DOWN.rotateVector(vec).toString());
+        assertEquals(new Vector3d(0.8, 0.1, 0.3).toString(), BlockFace.UP.rotateVector(vec).toString());
+        assertEquals(new Vector3d(0.9, 0.2, 0.7).toString(), BlockFace.WEST.rotateVector(vec).toString());
+        assertEquals(new Vector3d(0.7, 0.2, 0.1).toString(), BlockFace.SOUTH.rotateVector(vec).toString());
+        assertEquals(new Vector3d(0.3, 0.2, 0.9).toString(), BlockFace.NORTH.rotateVector(vec).toString());
     }
 }
